Push validation of Cmd validity into the same place as we perform Cmds
Previously we thought it would be nice to validate Cmds at the site of
generation, and only pass valid Cmds to be updated. Steve pointed out
that this would result in duplicated work, as we'd effectively perform
updates twice (once to check the validity of the resulting state, and a
second time to actually perform the Cmd).

This is lame, so instead we only return a new State if the Cmd was
successsfully applied.

Note that we're throwing away errors as far as the program is concerned
- the cause of any invalid Cmds will be logged instead. This might get
noisy, but we'll deal with that if it comes up.