Fix check for moving through regions to correctly check coordinates along a negative vector
Previously we were using ranges (unconditionally) to generate the region
between to coordinates, but sometimes the first coordinate passed is
greater than the second (along the axis which we are translating).

Now we both constrain the creation of LLDs so that invalid ones can't be
created, and support checking both positive and negative regions.

We also move all the Geometry related stuff into it's own module.